feat: refactor actor run response handling and improve data fetching logic
