+++ src/tools/default/get_actor_run.ts
@@ -1,9 +1,9 @@
 import type { InternalToolArgs, ToolEntry } from '../../types.js';
 import { logHttpError } from '../../utils/logging.js';
+import { fetchActorRunData } from '../core/actor_run_response.js';
 import {
     buildGetActorRunError,
     buildGetActorRunSuccessResponse,
-    fetchActorRunData,
     getActorRunArgs,
     getActorRunMetadata,
 } from '../core/get_actor_run_common.js';
